cleanup, entity stuff
changes entlist to a new class. Worked on map reloading. Began adding spawn functionality for entities

+++ input.cpp
@@ -33,8 +33,8 @@ void CursorMove(GLFWwindow* win, double xpos, double ypos)
 	xold = xpos;
 	yold = ypos;
 
-	in.yaw -= dxpos * sensitivity;
-	in.pitch += dypos * sensitivity;
+	in.yaw -= (float)(dxpos * sensitivity);
+	in.pitch += (float)(dypos * sensitivity);
 
 	if (in.yaw >= 360)
 		in.yaw -= 360;
